feat: 新增依赖操作可以重新安装删除的依赖；增加依赖切换

+++ src/views/home/components/dependence/tools.vue
@@ -10,7 +10,9 @@
       <i class="i-ep-setting" />
     </el-tooltip> -->
     <el-tooltip content="查看依赖主页">
-      <div class="b-rd-5px bg-gray-200 p-3px hover:bg-gray-400">
+      <div
+        class="b-rd-5px bg-primary bg-op-20 p-3px color-primary hover:bg-op-60"
+      >
         <div
           class="i-material-symbols:home-outline"
           @click="handleClickTools('home')"
@@ -18,23 +20,55 @@
       </div>
     </el-tooltip>
     <el-tooltip content="查看依赖关系">
-      <div class="b-rd-5px bg-gray-200 p-3px hover:bg-gray-400">
+      <div
+        class="b-rd-5px bg-primary bg-op-20 p-3px color-primary hover:bg-op-60"
+      >
         <div
           class="i-material-symbols:conversion-path"
           @click="handleClickTools('relation')"
         />
       </div>
     </el-tooltip>
-    <el-tooltip content="删除依赖">
-      <div
-        class="b-rd-5px bg-red-200 p-3px color-error hover:bg-red-400 hover:color-white"
+    <template v-if="!data.is_del">
+      <el-tooltip
+        :content="
+          data.is_dev ? '转换为dependencies依赖' : '转换为devDependencies依赖'
+        "
       >
         <div
-          class="i-material-symbols:delete-outline"
-          @click="handleClickTools('delete')"
-        />
-      </div>
-    </el-tooltip>
+          class="b-rd-5px bg-primary bg-op-20 p-3px color-primary hover:bg-op-60"
+        >
+          <div
+            class="i-material-symbols:transform"
+            @click="handleClickTools('transfer')"
+          />
+        </div>
+      </el-tooltip>
+    </template>
+    <template v-if="data.is_del">
+      <el-tooltip content="重新安装依赖，也可以选择下面最新版本进行安装">
+        <div
+          class="b-rd-5px bg-warning bg-op-20 p-3px color-warning hover:bg-op-60"
+        >
+          <div
+            class="i-material-symbols:restore-page-outline"
+            @click="handleClickTools('restore')"
+          />
+        </div>
+      </el-tooltip>
+    </template>
+    <template v-if="!data.is_del">
+      <el-tooltip content="删除依赖">
+        <div
+          class="b-rd-5px bg-red-200 p-3px color-error hover:bg-red-400 hover:color-white"
+        >
+          <div
+            class="i-material-symbols:delete-outline"
+            @click="handleClickTools('delete')"
+          />
+        </div>
+      </el-tooltip>
+    </template>
   </div>
 </template>
 <script setup>
@@ -50,6 +84,8 @@ const { data = {} } = defineProps({
 });
 const loadings = ref({
   delete: false,
+  restore: false,
+  transfer: false,
 });
 
 const root = ref(null);
@@ -62,6 +98,10 @@ const handleClickTools = (type) => {
       router.push({ path: `/${encodeURIComponent(data.name)}/relation` });
       break;
     case "delete":
+      if (loadings.value.delete) {
+        ElMessage.warning("正在删除，请稍后!");
+        return;
+      }
       loadings.value.delete = true;
       packageStore.deletePackage(data).then(
         () => {
@@ -70,6 +110,40 @@ const handleClickTools = (type) => {
         },
         () => {
           loadings.value.delete = false;
+        }
+      );
+      break;
+    case "restore":
+      if (loadings.value.restore) {
+        ElMessage.warning("正在恢复安装，请稍后!");
+        return;
+      }
+      // 恢复安装
+      loadings.value.restore = true;
+      packageStore.installPackage(data).then(
+        () => {
+          // 删除成功
+          loadings.value.restore = false;
+        },
+        () => {
+          loadings.value.restore = false;
+        }
+      );
+      break;
+    case "transfer":
+      if (loadings.value.transfer) {
+        ElMessage.warning("正在切换安装，请稍后!");
+        return;
+      }
+      // 恢复安装
+      loadings.value.transfer = true;
+      packageStore.installPackage({ ...data, is_change: true }).then(
+        () => {
+          // 删除成功
+          loadings.value.transfer = false;
+        },
+        () => {
+          loadings.value.transfer = false;
         }
       );
       break;
